server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
sqlmodel==0.0.22
httpx==0.27.2
pydantic-settings==2.6.0
pyyaml==6.0.2
pillow==11.0.0
pytest==8.3.3
python-multipart==0.0.12
sse-starlette==2.1.3
cryptography==43.0.3
requests==2.32.3
pypdf==5.0.1
